rpi4: Enable basic CPPC frequency scaling
This uses the arm mailbox to trigger secure interrupts
in EL3 which can then forward the requests to the videocore
firmware to adjust the core frequencies. As that mailbox
interface affects all the cores on the machine rather than
individually we need to add _PSD entries to put all
the cpus in the same powerstate domain.

Signed-off-by: Jeremy Linton <jeremy.linton@arm.com>