Platform/RaspberryPi/Acpi: Convert to ACPIv4 fan
ACPIv4 adds variable speed fans. The RPi POE hat
has one of these variable speed fans, and we can
control it if we trap to EL3 to forward a mailbox
commmand to the videocore firmware to change the
speed. Its possible we could try to control
the serial bus its hanging off ourselves, but
that is a heavier lift in AML than needed.

Rather than two different SSDT's for the two
fan profiles, lets upgrade the GPIO fan to
a variable speed fan as well, but lie about
the speed selection and only turn it on/off.

Signed-off-by: Jeremy Linton <jeremy.linton@arm.com>