Platform/RaspberryPi: Move GPIO/SPI/I2C to SSDT
The UEFI firmware uses the GPIO port for the fan and
real soon now the runtime SPI variable store. As such
we need to be able to either isolate those devices from
the OS or we risk clashing with OS's that reconfigure
the GPIO pins. Ideally we would just rip this out
and use _DSM() or just individual device power
on/off methods to adjust the GPIO pins when needed.

For now, lets leave it since windows at least knows
about it. In the future we will decide whether the
firmware is controlling something (SPI!) based on
whether the user has enabled the GPIO block.

Tested-by: Ard Biesheuvel <ardb@kernel.org>
Signed-off-by: Jeremy Linton <jeremy.linton@arm.com>