RaspberryPi: Add core I2C drivers
There are a number of I2C interfaces on the rpi,
some of which are available on the GPIO connector
and are utilized by various HATs. In particular we
are interrested in the RTCs (usually based on DS1037)
which are attached to GPIO2/3 which can be pin muxed
to I2C #1.

This commit adds a basic runtime utilizable I2C Dxe
and a platform driver which can bind it to the ds1307 rtc
driver.

Signed-off-by: Jeremy Linton <jeremy.linton@arm.com>